Review changes added

+++ src/main/java/com/example/taxibillingsystem/contract/response/UserResponse.java
@@ -14,6 +14,5 @@ public class UserResponse {
     private String name;
     private String email;
     private String password;
-    private long accountBalance;
+    private double accountBalance;
 }
-
