+++ src/main/java/com/example/taxibillingsystem/model/User.java
@@ -1,10 +1,10 @@
 package com.example.taxibillingsystem.model;
 
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -19,11 +19,10 @@ import lombok.NoArgsConstructor;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-
     private long userId;
+
     private String name;
     private String email;
     private String password;
-    private long accountBalance=0;
+    private double accountBalance = 0;
 }
-
